requirements.txt
numpy~=1.23.3
scipy~=1.9.1
yfinance~=0.2.12
python-dateutil~=2.8.2
matplotlib~=3.5.3